requirements.txt
flask
torch
torchvision
opencv-python
numpy
faiss-cpu
